Start with bare(r)-bones project
Deciding what to give to the students and what to have them implement
themselves was a bit difficult. A part of me wanted to throw the entire
project at them...but that felt like too much. Really, in this lab, we
should be focusing on the custom cells primarily. I took out what I
thought was relevant to that lesson, so the students can do it
themselves.